Delete HeartbeatAggregator and put its logic directly into the DistributedStreamService
