+++ cdap-data-fabric/src/main/java/co/cask/cdap/data/stream/service/LocalStreamService.java
@@ -32,8 +32,8 @@ import com.google.inject.Inject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicLong;
 
 /**
  * Stream service running in local mode.
@@ -45,8 +45,7 @@ public class LocalStreamService extends AbstractStreamService {
   private final StreamAdmin streamAdmin;
   private final StreamWriterSizeCollector streamWriterSizeCollector;
   private final StreamMetaStore streamMetaStore;
-  private final Map<String, Long> streamsInitCounts;
-  private final Map<String, Long> streamsBaseCounts;
+  private final Map<String, Aggregator> aggregators;
   private boolean isInit;
 
   @Inject
@@ -62,23 +61,16 @@ public class LocalStreamService extends AbstractStreamService {
     this.streamMetaStore = streamMetaStore;
     this.streamWriterSizeCollector = streamWriterSizeCollector;
     this.notificationService = notificationService;
-    this.streamsBaseCounts = Maps.newHashMap();
-    this.streamsInitCounts = Maps.newHashMap();
-    isInit = true;
+    this.aggregators = Maps.newConcurrentMap();
+    this.isInit = true;
   }
 
   @Override
   protected void initialize() throws Exception {
     for (StreamSpecification streamSpec : streamMetaStore.listStreams()) {
-      long filesSize = 0;
-      try {
-        StreamConfig config = streamAdmin.getConfig(streamSpec.getName());
-        filesSize = StreamUtils.fetchStreamFilesSize(config);
-      } catch (IOException e) {
-        LOG.error("Could not compute sizes of files for stream {}", streamSpec.getName());
-      }
-      streamsInitCounts.put(streamSpec.getName(), filesSize);
-      streamsBaseCounts.put(streamSpec.getName(), filesSize);
+      StreamConfig config = streamAdmin.getConfig(streamSpec.getName());
+      long filesSize = StreamUtils.fetchStreamFilesSize(config);
+      aggregators.put(streamSpec.getName(), new Aggregator(streamSpec.getName(), filesSize));
     }
   }
 
@@ -91,40 +83,62 @@ public class LocalStreamService extends AbstractStreamService {
   protected void runOneIteration() throws Exception {
     // Get stream size - which will be the entire size - and send a notification if the size is big enough
     for (StreamSpecification streamSpec : streamMetaStore.listStreams()) {
-      Long initSize = streamsInitCounts.get(streamSpec.getName());
-      if (initSize == null) {
+      Aggregator aggregator = aggregators.get(streamSpec.getName());
+      if (aggregator == null) {
         // First time that we see this Stream here
-        initSize = (long) 0;
-        streamsInitCounts.put(streamSpec.getName(), initSize);
-        streamsBaseCounts.put(streamSpec.getName(), initSize);
+        aggregator = new Aggregator(streamSpec.getName(), 0);
+        aggregators.put(streamSpec.getName(), aggregator);
       }
-      long absoluteSize = initSize + streamWriterSizeCollector.getTotalCollected(streamSpec.getName());
-
-      if (isInit || absoluteSize - streamsBaseCounts.get(streamSpec.getName()) >
-        Constants.Notification.Stream.DEFAULT_DATA_THRESHOLD) {
-        try {
-          publishNotification(streamSpec.getName(), absoluteSize);
-        } finally {
-          streamsBaseCounts.put(streamSpec.getName(), absoluteSize);
-        }
-      }
+      aggregator.checkAggregatedSize();
     }
     isInit = false;
   }
 
-  private void publishNotification(String streamName, long absoluteSize) {
-    NotificationFeed streamFeed = new NotificationFeed.Builder()
-      .setNamespace("default")
-      .setCategory(Constants.Notification.Stream.STREAM_FEED_CATEGORY)
-      .setName(streamName)
-      .build();
-    try {
-      notificationService.publish(streamFeed, new StreamSizeNotification(System.currentTimeMillis(), absoluteSize))
-        .get();
-    } catch (NotificationFeedException e) {
-      LOG.warn("Error with notification feed {}", streamFeed, e);
-    } catch (Throwable t) {
-      LOG.warn("Could not publish notification on feed {}", streamFeed.getId(), t);
+  /**
+   * Aggregate the sizes of all stream writers. A notification is published if the aggregated
+   * size is higher than a threshold.
+   */
+  private final class Aggregator {
+    private final AtomicLong streamInitSize;
+    private final NotificationFeed streamFeed;
+    private final String streamName;
+    private final AtomicLong streamBaseCount;
+
+    protected Aggregator(String streamName, long baseCount) {
+      this.streamName = streamName;
+      this.streamInitSize = new AtomicLong(baseCount);
+      this.streamBaseCount = new AtomicLong(baseCount);
+      this.streamFeed = new NotificationFeed.Builder()
+        .setNamespace(Constants.DEFAULT_NAMESPACE)
+        .setCategory(Constants.Notification.Stream.STREAM_FEED_CATEGORY)
+        .setName(streamName)
+        .build();
+    }
+
+    /**
+     * Check that the aggregated size of the heartbeats received by all Stream writers is higher than some threshold.
+     * If it is, we publish a notification.
+     */
+    public void checkAggregatedSize() {
+      long sum = streamInitSize.get() + streamWriterSizeCollector.getTotalCollected(streamName);
+      if (isInit || sum - streamBaseCount.get() > Constants.Notification.Stream.DEFAULT_DATA_THRESHOLD) {
+        try {
+          publishNotification(sum);
+        } finally {
+          streamBaseCount.set(sum);
+        }
+      }
+    }
+
+    private void publishNotification(long absoluteSize) {
+      try {
+        notificationService.publish(streamFeed, new StreamSizeNotification(System.currentTimeMillis(), absoluteSize))
+          .get();
+      } catch (NotificationFeedException e) {
+        LOG.warn("Error with notification feed {}", streamFeed, e);
+      } catch (Throwable t) {
+        LOG.warn("Could not publish notification on feed {}", streamFeed.getId(), t);
+      }
     }
   }
 }
